Inquiry and Quotation
Creating both backend and frontend architecture. few of features has been  installed.

Next development focusing on POST Inquiry, Inquiry Item, Quotation, Quotation Item.

Major problem was on making insertion feature both on inquiry and quotation items.

Developer must be concern of the interface that will be used by the user to input the data. Common solution is making modal interface for pick an items to be included in Inquiry and Quotation for make sure that the system will be failed and reducing the numbers of misleading input flow.

After this done we can take a big step for calculation of a project. in that case is making both layout and flow data is quite beauty.